Refactor forms to utilize updated Filament schema components, enhancing consistency across Coupon, Movie, Screen, Seat, and Theater resources. Add CountrySeeder to DatabaseSeeder for improved data seeding.
