+++ app/Filament/Resources/ScreenResource/Pages/Forms/ScreenResourceForm.php
@@ -3,8 +3,8 @@
 namespace App\Filament\Resources\ScreenResource\Pages\Forms;
 
 use App\Filament\Contracts\ResourceFieldContract;
-use Filament\Forms\Components\Grid;
-use Filament\Forms\Components\Section;
+use Filament\Schemas\Components\Grid;
+use Filament\Schemas\Components\Section;
 use Filament\Forms\Components\Select;
 use Filament\Forms\Components\TextInput;
