+++ app/Filament/Resources/TheaterResource/Pages/Forms/TheaterResourceForm.php
@@ -4,8 +4,8 @@ namespace App\Filament\Resources\TheaterResource\Pages\Forms;
 
 use App\Filament\Contracts\ResourceFieldContract;
 use App\Models\User;
-use Filament\Forms\Components\Grid;
-use Filament\Forms\Components\Section;
+use Filament\Schemas\Components\Grid;
+use Filament\Schemas\Components\Section;
 use Filament\Forms\Components\Select;
 use Filament\Forms\Components\Textarea;
 use Filament\Forms\Components\TextInput;
